add init_db, marginalized out from app, move extensions and config into core to make folder structure neater
