+++ backend/server/app.py
@@ -5,8 +5,8 @@ from dotenv import load_dotenv
 from flask import Flask, Response
 from flask_cors import CORS
 
-from server.config import Config
-from server.extensions import db, login_manager
+from server.core.config import Config
+from server.core.extensions import db, login_manager
 from server.models.user import User
 from server.routes.auth import auth_bp
 from server.routes.profile import profile_bp
@@ -14,7 +14,8 @@ from server.routes.route_service import route_service_bp
 from server.utils.errors import APIError, handle_api_error
 from server.utils.response import api_response
 
-load_dotenv()
+load_dotenv(".env.shared")
+load_dotenv(".env.local", override=True)
 
 app = Flask(__name__)
 app.config.from_object(Config)
@@ -54,9 +55,6 @@ app.register_blueprint(route_service_bp, url_prefix="/api")
 def home() -> tuple[Response, int]:
     return api_response(True, message="SmartRide Backend Running", status_code=200)
 
-
-with app.app_context():
-    db.create_all()
 
 if __name__ == "__main__":
     app.run(host="0.0.0.0", debug=True, port=5000)
