+++ backend/server/utils/ors_api.py
@@ -3,7 +3,8 @@ import os
 from dotenv import load_dotenv
 
 # Load .env for API key
-load_dotenv()
+load_dotenv(".env.shared")
+load_dotenv(".env.local", override=True)
 
 ors_api_key = os.getenv("ORS_API_KEY")
